refactor (ManifestJsonParser.js): relocate to package root
* update tests and paths
* relocate test spec to match pathing
