+++ lib/ManifestJsonParser.js
